Add Rather not say gender (#115)
People have complained that they want to add more genders than Male/Female. It makes sense. The quick workaround for this is to add a rather not say choice everywhere we mention the gender for people, which would be selected by default. In the future we'll be able to add more genders if needed.

This change requires migration.

Fix #44