Remove encryption in the journal table
I used to encrypt this but the decryption randomly generates problems sometimes. I wonder why. Is it because of special characters that are included and get corrupted while encrypted? I don't know, but I've found that removing the encryption (which is not necessary anyway), solves the problem - until we find a better solution.