Changes the unique constraint for contacts from 'email' to ['account_id', 'email']
I think the unique constraint is too eager for the contacts table. 

If a user wants to add a contact with an already existing email, they'll encounter an error. Since they aren't the one who added that email to their contacts, it will be confusing & frustrating. 

I think this is a bug and the constraint should be `['account_id', 'email']`.