Remove encryption on the Gifts table (#249)
This is just causing too much headache, especially since I made the rookie mistake of not checking the max length of some entries when inserting the gift. That leads to errors in the decryption process.

In production, out of the thousands of gifts already entered, only 5 are not decryptable. I have no other choice than resetting these 5 rows. I'm really sorry about this.

If you do end up having problems running the migration in your own instance, you need to remove the rows that are un-decryptable in the Gifts table as well.