Show reminders for the current day on the dashboard.
This delays the SetNextReminderDate job by 48 hours so that reminders'
next_expected_date does not change until after the current day has
passed for users in all timezones. This also changes the condition in
getRemindersForMonth() on the Account model so that reminders for the
current month will be shown from the start of the current day instead
of the current time.

This should not interfere with any recurring reminders because the
minimum interval for a reminder is 1 week.