Fix the store() function
Currently, attempts to use the POST /contactfields/ endpoint as outlined in the documentation at https://www.monicahq.com/api/contactfields#create-a-contact-field fail with a 404 error. This is because the store() function incorrectly calls validateUpdate, passing in the contact_field_type_id: the validation correctly fails, as that is NOT a valid contact_field_id. Removing that incorrect validation step fixes this error. There possibly should be a validation step to ensure that the contact_field_type_id is a valid value, but I'm assuming the DB's referential integrity check will take care of that for us.