Switch to HTTPS schema early
When using nginx as a https termination reverse proxy the assets and
URL in the html were generated as HTTP causing the browser not to
trust them, especially the login form.

This fixes the issue by switching to HTTPS early in the app startup
allowing laravel to generate URLs to have the HTTPS schema.

Issue #835