Documentation: Add instructions for configuring SES to send mail.
In Issue 860, there's a link to some documentation on how to use AWS SES for Email sending.

The current server installation instructions make no mention of how to configure an Email server, nor that it's necessary for the registation flow not to bug out.

This commit adds a new documentation page, linked to from the installation instructions, which give some examples of mail providers, and instructions on how to configure SES as a mail server.