docs: Removed minimum hardware requirements -- 3.0 cannot be built/run
Prior versions worked because yarn assets were pre-built and included.
With the removal of these, I don't think Monica can be successfully
updated on a machine with 768MB or less of RAM.  I think yarn's memory
requirements are too high.

Monica had been working fine on a Debian VPS with 256MB RAM.  It was a
little slow but fully functional.  While trying to upgrade to 3.0, I
could not successfully run the `yarn run production` command.  Despite
being left for 6+ hours on several attempts, it simply freezes and does
not go any further during the "fetching packages" stage.

I tried to build 3.0 on another Debian VPS with 768MB of RAM.  It
got further in the process but repeatedly failed with this error:

    error Command failed with exit code 134

Which, as far as I can tell from Google, is because it runs out of
memory.

Given that, I'm no longer comfortable recommending Monica run on such a
constrained machine.