style: Corrected 'birthdate' to 'birthday'
"Birthdate" is not used in English; it does not appear in the Shorter
Oxford English Dictionary.

In English, two forms are used: "birthday", which refers to the
celebration *and* the birth's date, and "date of birth", which is used
in more formal contexts.  Someone celebrates their birthday; someone
confirms their date of birth.