README change: referral to Asperger & Alzheimer's
This isn't code breaking! But I saw some phrasing that could perhaps be improved upon, to be more accommodating to as much people as possible.
This is probably not important to a lot of people nor general knowledge, so some context behind the rephrasing:
- "Suffer from" is a phrasing that got co-opted by people who see autism as a bad thing, something that has to be cured
- Asperger's syndrome, a.k.a. "high functioning autism" is only one subset of the spectrum. It makes sense that people with AS would be the most likely to reach out and give feedback, it'd be nice to also acknowledge the other subsets, along with "social disorders like" to include stuff that is akin to, but not exactly, ASD
- "Affected by Alzheimer's" also includes people who have someone with Alzheimer's in their life, who then also get use out of this app to make notes about their relative/friend/etc!

Hopefully these changes make sense, feel free to ask any questions, and thank you for reading!